+++ src/WebRoutes.tsx
@@ -57,6 +57,9 @@ import IndexDetailDronerPoint from "./pages/droner/IndexDetailDronerPoint";
 import EditCampaignPoint from "./pages/campaign/point/EditCampaignPoint";
 import ConditionDroner from "./pages/setting/pointSetting/ConditionDroner";
 import ConditionFarmer from "./pages/setting/pointSetting/ConditionFarmer";
+import IndexReceivePoint from "./pages/point/IndexReceivePoint";
+import IndexPlanningPoint from "./pages/point/IndexPlanningPoint";
+import DetailReceivePoint from "./pages/point/DetailReceivePoint";
 
 const WebRoutes: React.FC<any> = () => {
   return (
@@ -141,6 +144,12 @@ const WebRoutes: React.FC<any> = () => {
           />
           <Route path="/ConditionDroner" element={<ConditionDroner />} />
           <Route path="/ConditionFarmer" element={<ConditionFarmer />} />
+          <Route path="/IndexReceivePoint" element={<IndexReceivePoint />} />
+          <Route path="/IndexPlanningPoint" element={<IndexPlanningPoint />} />
+          <Route
+            path="/DetailReceivePoint/:id"
+            element={<DetailReceivePoint />}
+          />
         </Route>
 
         <Route element={<PublicRoute />}>