+++ src/datasource/TaskDatasource.ts
@@ -70,7 +70,6 @@ export class TaskDatasource {
       });
   }
   static updateNewTask(data: UpdateNewTask): Promise<any> {
-    console.log("data", data);
     return httpClient
       .patch(BASE_URL + "/tasks/task/" + data.id, data)
       .then((response) => {