+++ src/entities/CampaignPointEntites.ts
@@ -24,7 +24,8 @@ export interface CampaignEntiry {
   application: string;
   createAt: string;
   updateAt: string;
-  isDelete: boolean;
+  isDeleteFarmer: boolean;
+  isDeleteDroner: boolean;
 }
 
 export const CampaignEntiry_INIT: CampaignEntiry = {
@@ -40,7 +41,8 @@ export const CampaignEntiry_INIT: CampaignEntiry = {
   application: "",
   createAt: "",
   updateAt: "",
-  isDelete: false,
+  isDeleteFarmer: false,
+  isDeleteDroner: false,
 };
 
 export interface CampaignListEntity {