+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## [Version - 0]
+## Fix [v0.28.2] 2023-05-22
+- fixed discount coupon in new task
+## Fix [v0.28.1] 2023-05-22
+- fixed campoign point entity
+## Feature [v0.28.0] 2023-05-19
+- added planning point menu
 ## Fix [v0.27.4] 2023-05-17
 - fixed logic delete & edit in campaign point menu
 ## Fix [v0.27.3] 2023-05-17